+++ src/main/java/io/github/poshjosh/ratelimiter/raas/services/PermitService.java
@@ -26,32 +26,28 @@ public class PermitService {
     }
 
     @Async
-    public void tryAcquireAsync(String rateId, int permits, HttpRequestDto httpRequestDto) {
+    public void tryAcquireAsync(String rateId, int permits, /* @Nullable */ HttpRequestDto httpRequestDto) {
         tryAcquire(rateId, permits, httpRequestDto);
     }
 
-    public boolean tryAcquire(String rateId, int permits, HttpRequestDto httpRequestDto) {
-        addRateIdHeaderValue(httpRequestDto, rateId);
-        final RequestInfo requestInfo = httpRequestMapper.toRequestInfo(httpRequestDto);
+    public boolean tryAcquire(String rateId, int permits, /* @Nullable */ HttpRequestDto httpRequestDto) {
+        final RequestInfo requestInfo = httpRequestMapper
+                .toRequestInfo(httpRequestDto, getAdditionalHeaders(rateId));
         final boolean acquired = rateLimiterRegistry.tryAcquire(requestInfo, permits);
         log.debug("Acquired {}, {} permits from rate: {} for {}",
                 acquired, permits, rateId, httpRequestDto);
         return acquired;
     }
 
-    public boolean isAvailable(String rateId, HttpRequestDto httpRequestDto) {
-        addRateIdHeaderValue(httpRequestDto, rateId);
-        final RequestInfo requestInfo = httpRequestMapper.toRequestInfo(httpRequestDto);
+    public boolean isAvailable(String rateId, /* @Nullable */ HttpRequestDto httpRequestDto) {
+        final RequestInfo requestInfo = httpRequestMapper
+                .toRequestInfo(httpRequestDto, getAdditionalHeaders(rateId));
         final boolean available = rateLimiterRegistry.isWithinLimit(requestInfo);
         log.debug("Permit available {}, rate: {}, for {}", available, rateId, httpRequestDto);
         return available;
     }
 
-    private void addRateIdHeaderValue(HttpRequestDto httpRequestDto, String value) {
-        Map<String, List<String>> headers = httpRequestDto.getHeaders() == null
-                || httpRequestDto.getHeaders().isEmpty()
-                ? new HashMap<>() : new HashMap<>(httpRequestDto.getHeaders());
-        headers.put(RateLimiterConfiguration.RATE_ID_HEADER, List.of(value));
-        httpRequestDto.setHeaders(headers);
+    private Map<String, List<String>> getAdditionalHeaders(String rateId) {
+        return Map.of(RateLimiterConfiguration.RATE_ID_HEADER, List.of(rateId));
     }
 }